refactor [service,controller,VO,DTO,mapper,impl]: 代码优化

+++ meeting-appointment-modules/src/main/java/com/jfzt/meeting/service/MeetingDeviceService.java
@@ -51,5 +51,5 @@ public interface MeetingDeviceService extends IService<MeetingDevice> {
      * @param id 设备id
      * @return 成功信息
      */
-    Result<Object> updateStatusDevice(Integer id);
+    Result<Object> updateStatusDevice(Long id);
 }
